requirements.txt
# Web framework
fastapi==0.95.2
uvicorn[standard]==0.22.0
flask==2.3.3

# Database
sqlalchemy==2.0.19
databases[postgresql]==0.9.0
asyncpg==0.29.0

# Message Queue (Azure Service Bus)
azure-servicebus==7.11.1

# Environment variables
python-dotenv==1.0.0

# GraphQL API
strawberry-graphql==0.141.0

# Caching
fastapi-cache2==0.2.2

# Metrics
prometheus-fastapi-instrumentator==6.1.0

# Logging (structured/log file or queue)
structlog==23.1.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.0
httpx==0.25.0

# Encription
bcrypt==4.1.2
cryptography==42.0.5

# For ASGI server with FastAPI + Gunicorn + Uvicorn worker
uvicorn==0.22.0
gunicorn==23.0.0